+++ torch/optim/__init__.py
@@ -8,7 +8,7 @@ future.
 from .adadelta import Adadelta
 from .adagrad import Adagrad
 from .adam import Adam
-from .adamw import AdamW
+from .optimizerw import AdamW
 from .sparse_adam import SparseAdam
 from .adamax import Adamax
 from .asgd import ASGD
@@ -22,7 +22,7 @@ from . import lr_scheduler
 del adadelta
 del adagrad
 del adam
-del adamw
+del optimizerw
 del sparse_adam
 del adamax
 del asgd
